Dev canh

See merge request shsoft/mini-erp-team2!24

+++ src/main/java/com/shsoftvina/erpshsoftvina/service/impl/NotificationImpl.java
@@ -1,5 +1,6 @@
 package com.shsoftvina.erpshsoftvina.service.impl;
 
+import com.shsoftvina.erpshsoftvina.constant.ApplicationConstant;
 import com.shsoftvina.erpshsoftvina.constant.NotificationConstant;
 import com.shsoftvina.erpshsoftvina.converter.NotificationConverter;
 
@@ -13,6 +14,7 @@ import com.shsoftvina.erpshsoftvina.model.response.notification.NotificationComm
 import com.shsoftvina.erpshsoftvina.model.response.notification.NotificationResponse;
 import com.shsoftvina.erpshsoftvina.service.NotificationService;
 import com.shsoftvina.erpshsoftvina.utils.FileUtils;
+import com.shsoftvina.erpshsoftvina.utils.MessageErrorUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -38,30 +40,24 @@ public class NotificationImpl implements NotificationService {
     }
 
     @Override
-    public NotificationResponse findById(String id) {
-        Notification notification = notificationMapper.findById(id);
-        return notificationConverter.toResponse(notification);
-    }
-
-    @Override
     public int createNoti(CreateNotificationRequest createNotificationRequest) {
 
-        if (createNotificationRequest.getFile().length > NotificationConstant.NUMBER_FILE_LIMIT) {
-            throw new FileTooLimitedException("Max file is 3");
+        if (createNotificationRequest.getFile().length > ApplicationConstant.NUMBER_UPLOAD_FILE_LIMIT) {
+            throw new FileTooLimitedException(MessageErrorUtils.notAllowFileSize());
         }
 
         String dir = NotificationConstant.UPLOAD_FILE_DIR;
 
-        List listFileNameSaveFileSuccess = FileUtils.saveMultipleFilesToServer(request, dir, createNotificationRequest.getFile());
+        List<String> listFileNameSaveFileSuccess = FileUtils.saveMultipleFilesToServer(request, dir, createNotificationRequest.getFile());
         if(listFileNameSaveFileSuccess != null){
-            Notification notification = notificationConverter.toEntity(createNotificationRequest);
+            Notification notification = notificationConverter.toEntity(createNotificationRequest, listFileNameSaveFileSuccess);
             try{
                 notificationMapper.createNoti(notification);
+                return 1;
             } catch (Exception e){
                 FileUtils.deleteMultipleFilesToServer(request,dir, notification.getFile());
                 return 0;
             }
-            return 1;
         }
         return 0;
     }
@@ -69,25 +65,25 @@ public class NotificationImpl implements NotificationService {
     @Override
     public int updateNoti(UpdateNotificationRequest updateNotificationRequest, String id) {
 
-        if (updateNotificationRequest.getFile().length > NotificationConstant.NUMBER_FILE_LIMIT) {
-            throw new FileTooLimitedException("Max file is 3");
+        if (updateNotificationRequest.getFile().length > ApplicationConstant.NUMBER_UPLOAD_FILE_LIMIT) {
+            throw new FileTooLimitedException(MessageErrorUtils.notAllowFileSize());
         }
 
         Notification notification = notificationMapper.findById(id);
 
         if(notification == null){
-            throw new NotFoundException("id not found");
+            throw new NotFoundException(MessageErrorUtils.notFound("Id"));
         }
 
         String listFileNameOld = notification.getFile();
 
         String dir = NotificationConstant.UPLOAD_FILE_DIR;
 
-        List listFileNameSaveFileSuccess = FileUtils.saveMultipleFilesToServer(request, dir, updateNotificationRequest.getFile());
+        List<String> listFileNameSaveFileSuccess = FileUtils.saveMultipleFilesToServer(request, dir, updateNotificationRequest.getFile());
 
         if(listFileNameSaveFileSuccess != null){
             try{
-                notification = notificationConverter.toEntity(updateNotificationRequest, id);
+                notification = notificationConverter.toEntity(updateNotificationRequest, id, listFileNameSaveFileSuccess);
                 notificationMapper.updateNoti(notification);
                 return 1;
             } catch (Exception e){
@@ -101,15 +97,22 @@ public class NotificationImpl implements NotificationService {
     @Override
     public boolean delNoti(String id) {
         Notification notification = notificationMapper.findById(id);
-        notificationMapper.delNoti(id);
-        String dir = NotificationConstant.UPLOAD_FILE_DIR;
-        FileUtils.deleteMultipleFilesToServer(request, dir, notification.getFile());
-        return true;
+        if(notification == null){
+            throw new NotFoundException(MessageErrorUtils.notFound("Id"));
+        }
+        try {
+            notificationMapper.delNoti(id);
+            String dir = NotificationConstant.UPLOAD_FILE_DIR;
+            FileUtils.deleteMultipleFilesToServer(request, dir, notification.getFile());
+            return true;
+        } catch (Exception e){
+        }
+        return false;
     }
 
     @Override
-    public NotificationCommentsListResponse getCommentsByNotificationId(String notificationId){
-        Notification comments = notificationMapper.getCommentsByNotificationId(notificationId);
+    public NotificationCommentsListResponse findById(String id){
+        Notification comments = notificationMapper.findById(id);
         System.out.println(comments);
         return notificationConverter.toListCommentsResponse(comments);
     }
